memory/requirements.txt
crewai
crewai-tools
mem0ai
openai
python-dotenv
langchain
langchain-openai
langchain-core
# Zep dependencies
zep-cloud
langgraph
ipywidgets
